Added shared jest and ts-jest configuration
