+++ packages/ilib-common-config/tsconfig.json
@@ -1,29 +1,6 @@
 {
-    "extends": "@tsconfig/node12",
+    "extends": "./tsconfig.base.json",
     "compilerOptions": {
-        "target": "ES5",
-        "module": "CommonJS",
-        "moduleResolution": "node",
-        "types": ["jest", "node"],
-        "strict": true,
-        "esModuleInterop": true,
-        "skipLibCheck": true,
-        "forceConsistentCasingInFileNames": true,
-        "declaration": true,
-        "declarationMap": true,
-        "sourceMap": true,
-        "removeComments": false,
-        "noImplicitAny": true,
-        "noImplicitReturns": true,
-        "noImplicitThis": true,
-        "noUnusedLocals": false,
-        "noUnusedParameters": false,
-        "exactOptionalPropertyTypes": true,
-        "noImplicitOverride": true,
-        "noPropertyAccessFromIndexSignature": true,
-        "noUncheckedIndexedAccess": true,
-        "allowUnusedLabels": false,
-        "allowUnreachableCode": false,
         "outDir": "./lib",
         "rootDir": "./src"
     },
@@ -35,4 +12,4 @@
         "lib",
         "test"
     ]
-} 
+}
